requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
cffi==2.0.0
cryptography==46.0.5
fastapi==0.135.1
idna==3.11
invoke==2.2.1
markdown-it-py==4.0.0
mdurl==0.1.2
mysql-connector==2.2.9
netmiko==4.6.0
ntc_templates==9.0.0
paramiko==4.0.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyNaCl==1.6.2
pyserial==3.5
PyYAML==6.0.3
rich==14.3.3
ruamel.yaml==0.19.1
scp==0.15.0
starlette==0.52.1
textfsm==2.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
